refactor: remove planned.py, move stubs to their command files
- Move dataset stubs to dataset_commands.py (ready for Tier 2 implementation)
- Inline monitor/trace/model/agent stubs in app.py (1-2 commands each)
- Delete planned.py — no more catch-all stub file
